Use ShadowPicker for ResourcesImpl and ThemeImpl.
