+++ shadows/framework/src/main/java/org/robolectric/shadows/ShadowResources.java
@@ -8,14 +8,12 @@ import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
-import android.os.Build;
 import android.util.DisplayMetrics;
 import android.util.LongSparseArray;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.List;
-import org.robolectric.RuntimeEnvironment;
 import org.robolectric.annotation.Implementation;
 import org.robolectric.annotation.Implements;
 import org.robolectric.annotation.RealObject;
@@ -31,11 +29,7 @@ abstract public class ShadowResources {
       if (useLegacy()) {
         return ShadowLegacyResources.class;
       } else {
-        if (RuntimeEnvironment.getApiLevel() >= Build.VERSION_CODES.P) {
-          return null;
-        } else {
-          return ShadowArscResources.class;
-        }
+        return ShadowArscResources.class;
       }
     }
   }
